+++ ThirdPartyAdapters/inmobi/CHANGELOG.md
@@ -1,5 +1,13 @@
 ## InMobi Android Mediation Adapter Changelog
 
+#### Version 10.0.2.0 (In Progress)
+- Verified compatibility with InMobi SDK version 10.0.2.
+- Updated the minimum required Google Mobile Ads SDK version to 20.5.0.
+
+Built and tested with:
+- Google Mobile Ads SDK version 20.5.0.
+- InMobi SDK version 10.0.2.
+
 #### Version 10.0.1.0
 - Verified compatibility with InMobi SDK version 10.0.1.
 - Updated the minimum required Google Mobile Ads SDK version to 20.4.0.